cache: Move version reading code to cache package
The code to get the current version of a binary is currently located in
the various preflight files. Since for a given binary, the version we
expect, and how to get the binary version does not need to vary between
cache instances, we can consolidate these in the cache package.
